requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-tailwind==3.8.0
django-browser-reload==1.12.1
nltk==3.8.1